updated stuff to follow database schema

+++ projeto-ies/src/main/java/pt/deti/ua/services/EventService.java
@@ -23,7 +23,7 @@ public class EventService {
     }
 
     //fazer handle das exceptions depois ---> esta primeira é só para não dar erros enquanto não criarmos as classes de exceptions
-    public Event getEventById(Long id) {
+    public Event getEventById(int id) {
         Optional<Event> optionalEvent = eventRepository.findById(id);
         return optionalEvent.orElse(null);
     }
